Split up v6 configs
