+++ src/board/mod.rs
@@ -1,11 +1,17 @@
 #[cfg_attr(feature = "hw_v1", path = "hardware/v1.rs")]
 #[cfg_attr(feature = "hw_v2", path = "hardware/v2.rs")]
 #[cfg_attr(feature = "hw_v4", path = "hardware/v4.rs")]
-#[cfg_attr(feature = "hw_v6", path = "hardware/v6.rs")]
-#[cfg_attr( // We default to hw_v6 if no feature is selected to help rust-analyzer for example
-    // TODO
-    not(any(feature = "hw_v1", feature = "hw_v2", feature = "hw_v4", feature = "hw_v6")),
-    path = "hardware/v6.rs"
+#[cfg_attr(all(feature = "hw_v6", feature = "esp32s3"), path = "hardware/v6s3.rs")]
+#[cfg_attr(all(feature = "hw_v6", feature = "esp32c6"), path = "hardware/v6c6.rs")]
+#[cfg_attr( // We default to hw_v6/esp32c6 if no feature is selected to help rust-analyzer for example
+    not(any(
+        feature = "hw_v1",
+        feature = "hw_v2",
+        feature = "hw_v4",
+        all(feature = "hw_v6", feature = "esp32s3"),
+        all(feature = "hw_v6", feature = "esp32c6")
+    )),
+    path = "hardware/v6c6.rs"
 )]
 pub mod hardware;
